==> ext_code/r_code/r_markdowns/11_scRNA_TF_analysis.Rmd <==
---
title: 'Smed Cisreg 07.b: Transcription Factor Analysis'
author: 'Alberto Perez Posada @apposada'
date: '10/25/2023'
output:
  html_document: default
  pdf_document: default
editor_options: 
  markdown: 
    wrap: sentence
---

```{r setup, include=FALSE}

dir <- '/mnt/sda/alberto/projects/smed_cisreg/'
fcha <- function(){ gsub("-","",Sys.Date()) }

knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = dir)
options(scipen=999)

```

## Loading Necessary Packages

```{r warning = FALSE, message=FALSE}
library(ComplexHeatmap)
library(circlize)
library(colorspace)
library(ggplot2)
library(effects)
```

```{r}
# add sourcefolder here
source("code/r_code/functions/sourcefolder.R")
sourceFolder(folder="code/r_code/functions/",recursive = TRUE)
```

## Data Load

We start with our previous session.

```{r load_data}
load(
  "outputs/rda/smed_counts.rda"
)
load(
  "outputs/rda/smed_wgcna.rda"
)
```

We will also load the TF annotation of S. mediterranea

```{r load_TF_annotation}
smed_tfs <- read.delim2(
  "~/../Standard_References/Smed_Rink/20231127_Rosetta_EXTENDED.tsv",
  header = T)[,c("gene","Class")] #load column of plei ID and TF class
smed_tfs <- smed_tfs[smed_tfs$Class != "-",]
colnames(smed_tfs) <- c("id","class")
head(smed_tfs)
```

To this table we will add all the TFs that were found in the paper of Neiro et al., 2022

```{r}
smed_tfs_all <- read.delim2(
  "~/../Standard_References/Smed_Rink/20231127_Rosetta_EXTENDED.tsv",
  header = T)[,c(1,33:48)]
```

We subset the gene expression pseudo-bulk matrix to retrieve expression from the TFs.

```{r}
smed_tfs_cw <-
  smed_counts_norm_cw[
    rownames(smed_counts_norm_cw) %in% smed_tfs$id,
    ]
```

We can browse the expression level of different TFs using this function.

```{r}
plot_tf_smed <- function(x){
  if(x %in% rownames(smed_tfs_cw)) {
    barplot(
      height=unlist(c(
        smed_tfs_cw[
          grep(
            paste("^",x,"$",sep=""),
            rownames(smed_tfs_cw),
          ),
        ]
      )),
      col = smed_ctypes$col[match(colnames(smed_tfs_cw),smed_ctypes$ctype)],
      border = "#2F2F2F",
      las=2,
      cex.names=0.7,
      main= paste(
        x,
        " (",
        smed_tfs[grep(x,smed_tfs$id),2],
        ")\n",
        sep=""
      ),
      ylab="counts per million per cluster"
    )} else {
      stop("Name not in list of TFs.")
    }
}

```

For example, the hnf4 gene;

```{r}
hnf4 <- "h1SMcG0019688"
plot_tf_smed(hnf4)
```

As we have expression data of many transcription factors, we can visualise the global patterns of expression using heatmaps.
We will do so by scaling the log-transformed expression of TFs to obtain a z-score.

```{r}
smed_tfs_genecol <-
  data.frame(
    id = rownames(smed_tfs_cw),
    ctype = apply(
      smed_tfs_cw,
      1,
      highest_val # a custom function that tells which is the highest value
    )
  )

smed_tfs_genecol$ctype <- 
  factor(smed_tfs_genecol$ctype,levels = colnames(smed_tfs_cw))
smed_tfs_genecol <- 
  smed_tfs_genecol[order(smed_tfs_genecol$ctype),]

smed_tfs_fc <- smed_tfs_cw[match(smed_tfs_genecol$id,rownames(smed_tfs_cw)),]
```

And using the ComplexHeatmap package:

```{r, fig.height = 8, fig.width = 4}
col_smed_tfs_expr <- colorRamp2(
  c(0:3),
  colorRampPalette(c("#f1f5ff","#b1b8c4","#2c2b46"))(4)
  )

clu_ha = HeatmapAnnotation(
  name = "cell types",
  cluster = colnames(smed_tfs_fc),
  col = list( cluster = setNames(smed_ctypes$col[match(colnames(smed_tfs_fc),smed_ctypes$ctype)],colnames(smed_tfs_fc)))
  )
clu_ha@anno_list$cluster@show_legend <- FALSE
clu_ha@anno_list$cluster@label <- NULL

h1 <- Heatmap(
  name="z-score",
  t(scale(t(smed_tfs_fc))),
  col = col_smed_tfs_expr,
  show_row_names = FALSE,
  show_column_names = FALSE,
  cluster_rows=FALSE,
  cluster_columns=F,
  top_annotation = clu_ha,
  # right_annotation = smed_tfs_row_anno,
  row_title=NULL
)
draw(h1)
```

```{r}
pdf("graphics/tfs_fc_all_supp.pdf", height = 7, width = 3)
draw(h1)
dev.off()
```

```{r echo = FALSE}
# do we need any of this anymore??
# draw(smed_cor_hm)
# draw(smed_expr_zsco_hm)
# 
# pdf("graphics/smed_TFs_heatmap.pdf",width = 6, height = 8)
# draw(smed_expr_zsco_hm)
# dev.off()
```

## Analysing TFs and module connectivity

From the definition in the original WGCNA paper, the eigengene of a given module can be understood as:
  "The first principal component of a given module. It can be considereded a representative of the expression profiles of the genes in that given module." (slightly adapted for clarity)

For each gene, WGCNA defines a "fuzzy" measure of module membership by correlating the expression profile to that of the module eigengenes. If this value is closer to 1 it indicates that that gene is connected to many genes of that module.

We will aggregate the average expression profiles to use as eigengenes.

We can calculate the connectivity by correlating the average module expression profiles with the expression of TFs:

```{r}
tf_eigen <- 
  WGCNA::signedKME(
    scale(t(smed_tfs_cw)), # all tfs, not only those with CV > 1.25 as in wgcna markdown
    MEs, outputColumnName = ""
    )

min_kme <- .55

filt_top <- 
  apply(
    tf_eigen, 1,
    function(x){
      if(any(x > min_kme)){res = TRUE} else {res = FALSE} 
      return(res)
    }
  )

tf_eigen <- tf_eigen[filt_top,]

smed_tfs_kme <-
  data.frame(
    id = rownames(tf_eigen),
    module = apply(
      tf_eigen,
      1,
      highest_val_0 # a custom function that tells which is the highest value
    )
  )
smed_tfs_kme$module <- factor(smed_tfs_kme$module, levels = levels(smed_id_module$module))

smed_tfs_kme <- smed_tfs_kme[order(smed_tfs_kme$module),]

tf_eigen <- tf_eigen[
  match(smed_tfs_kme$id,rownames(tf_eigen))
  ,
  ]
```

And again, we can visualise using ComplexHeatmap

```{r, fig.height=8, fig.width=4, message = FALSE, warning = FALSE}
col_kme <- 
  colorRamp2(seq(0.3,0.8,len=10),colorRampPalette(rev(viridis_pastel))(10))

modules_ha <-
  HeatmapAnnotation(
    stacked = anno_barplot(
      hm_bp,
      gp = gpar(fill = smed_ctypes$col[match(colnames(hm_bp),smed_ctypes$ctype)],col=NA),
      border = FALSE,
      bar_width = 1
    ),
    annotation_name_side='right',
    gap = unit(5,"pt"),
    show_legend = FALSE
  )

h2 <- Heatmap(
  name="kME",
  tf_eigen,
  col=col_kme,
  show_row_names = FALSE,
  show_column_names = TRUE,
  cluster_rows=FALSE,
  cluster_columns=FALSE,
  top_annotation = modules_ha,
  column_names_side = "top",
  column_names_gp = gpar(fontsize = 8)
)

draw(h2)
```


```{r}
png("graphics/tfs_connectivity_all_supp.png", width = 512, height = 1280)
draw(h2)
dev.off()
```


## Common plot, plus TFs from the literature

We identified several TFs previously described in the literature whose region of expression within the animal is corroborated by our analyses.

```{r}
tfs_fig2 <- read.delim2("outputs/functional_annotation/tfs_fig2.tsv", header = TRUE)
head(tfs_fig2)
```

We will plot a conjoined heatmap of expression and connectivity, highlighting the position of these TFs in the figure.

```{r, fig.height = 8, fig.width = 6, warning = FALSE, message = FALSE}
tfs_fc_common <- t(scale(t(smed_tfs_fc)))
tfs_fc_common <- tfs_fc_common[rownames(smed_tfs_fc) %in% rownames(tf_eigen),]
tfs_eigen_common <- tf_eigen[rownames(tf_eigen) %in% rownames(tfs_fc_common),]
tfs_fc_common <- tfs_fc_common[match(rownames(tfs_eigen_common),rownames(tfs_fc_common)),]
#tfs_eigen_common <- tfs_eigen_common[match(rownames(tfs_fc_common),rownames(tfs_eigen_common)),]

where_tfs_common <- unlist(sapply(tfs_fig2$gene,function(x){grep(x,rownames(tfs_eigen_common))}))
tfs_rowanno_common <-
  rowAnnotation(
    TF = anno_mark(
      at = where_tfs_common,
      labels = translate_ids(names(where_tfs_common),tfs_fig2[,c(1,3)]))
  )

h1_main <- Heatmap(
  name="FC",
  tfs_fc_common,
  col=col_smed_tfs_expr,
  show_row_names = FALSE,
  show_column_names = FALSE,
  cluster_rows=FALSE,
  cluster_columns=F,
  top_annotation = clu_ha,
  right_annotation = tfs_rowanno_common,
  bottom_annotation = clu_ha,
  column_names_side = "bottom",
  row_title=NULL
)

h2_main <- Heatmap(
  name="kME",
  tfs_eigen_common,
  col=col_kme,
  show_row_names = FALSE,
  cluster_rows=FALSE,
  cluster_columns=F,
  top_annotation = modules_ha,
  # column_labels = smed_modules_table$codename,
  column_names_side = "top",
  column_names_gp = gpar(fontsize = 7)
)

draw(h2_main+h1_main)
# draw(h1_main+h2_main)
```


```{r}
pdf("graphics/tfs_expr_and_connectivity_main.pdf", width = 8, height = 10)
draw(h2_main+h1_main)
dev.off()
```

## Saving the data

We will save the important bits for further analysis in the rest of markdowns.

```{r}
save(
  # gene expression data
  smed_tfs_cw,
  smed_tfs_fc,
  # tf data
  smed_tfs,
  # neiro_tfs,
  tfs_fig2,
  # kME
  tf_eigen,
  tfs_eigen_common,
  # visual annotations
  modules_ha,
  col_kme,
  hm_bp,
  # ctypes_rowAnno,
  # clu_ha,
  # modules_ha,
  # wg_ha,
  #pick color palette for TFs
  # destination
  file = paste0(
    "./outputs/rda/",
    "tf_analysis.rda"
  )
)
```
